+++ src/lua/CommunityBalanceMod/Minimap/Armory_Server.lua
@@ -6,7 +6,6 @@ function Armory:OnResearchComplete(researchId)
     oldArmoryOnResearchComplete(self, researchId)
 
     if researchId == kTechId.AdvancedArmoryUpgrade then
-
         self.advanced = true
         if HasMixin(self, "MapBlip") then 
             self:MarkBlipDirty()