+++ docs/changelog.md
@@ -65,7 +65,7 @@
   - Dropping welders cost 2 tres (from 3 tres)
   - Autopickup for welders reduced from 5 to 1 second
   - ARCs dont deal damage to other ARCS anymore
-  - Grenade launchers dont inflict selfdamage anymore
+  - Selfdamage reduced by 66% (grenades/mines)
   - Bile damage accelerates weapon expiration
     - 1 Bile ~ 5 seconds
 
@@ -80,7 +80,6 @@
   - Added Icon for Drifter Eggs
   - Added Icon for deployed ARCS
   - Added Icon for Advanced Armory
-  - Marine structures under construction or powerless are slightly brighter
 
   - Alien Commander is able to see parasited mines